2024 Quest 10
Part 3 was a bit tricky (and I don't quite think my solution is general - I just stopped once it worked). Initially I missed the bit about unsolvable tiles (and was very confused when there were two Bs in a column). I think the bit that fixed it was eliminating unsolvable tiles (using @ since that's worth 0) so that they didn't pollute question marks on the wall. I might clean it up a little and actually use the calculated question marks