2024 Quest 13 Part 3
For parts 1 and 2, I started with a BFS, but that was incredibly slow and memory intensive, so I resorted to a (much faster) iterative solution for parts 2 and 3 (part 1 was "solved" by submitting a few potential times that were printed as partial solutions). For part 3, I started by iterating through all the starting values, then realised that was stupidly slow and started from the single end value (since paths are equivalent both ways), then found the minimum cost out of all the start values. This should not have taken as long as it did - maybe I should remember my iterative path finder in the future