2024 Quest 13
Slight reorder and refactor to solve all three parts. All three now build paths from the end (since it doesn't matter for parts 1 and 2, and it really matters for part 3). On my machine, it takes about 6 seconds with PyPy or 18 with CPython, and it doesn't use 97% of my RAM like the initial queue-based attempts did