Visualise 2024 Quest 14 in Minecraft
Generate an mcfunction file that will produce an oak tree where the player's head is. The logs even face the direction they were grown in (since it looked bad when they didn't). There are even leaves at the end! This will not work as is unless your name is "allan" and you have a Minecraft world called "Everybody Codes 2024" with the appropriate folders. You'll need to change the datapack path if you want to use this