Visualise 2024 Quest 15 Part 3 in Minecraft
It doesn't actually do the pathfinding (that came from the creator, though I probably could generate it. This is part of the reason I haven't done parts 1 & 2 yet). I just picked 15 flowers that I could find the names for. It generates so the start is where the player is standing (since that makes more sense that starting with a corner). It's possible to follow the path all the way through by always turning left (or right - as long as it's consistent)