2024 Quest 17
I could immediately tell that this was a minimum spanning tree problem, but there was no way I was writing that from scratch quickly so I used NetworkX. For part 1, I missed the bit about adding the number of stars to the size. For part 3 when I was manually evaluating the product, I accidentally cut off a digit in the multiplication.