2024 Quest 17 Part 3 Clustering Visualisation
Hue represents different regions. Value (brightness) represents the cluster size. The offset -65 was chosen so the duck ended up yellow for my input. There is almost certainly a better way to choose different hues so that disjoint groups don't appear too similar. The scaling of 3W and 5H is to improve the proportions of the duck