2024 Quest 18
For parts 1 and 2, it does a floodfill(?) from the start(s) to find the minimum cost for each space, then takes the largest palm cost. For part 3, it works backwards from every palm (since there are much fewer palms than dots) and finds the smallest sum for each point. Part 3 is still quite slow (~5 seconds) since it needs to do 189 fills and 12311 sums of 189 values (but that's a lot better than 12311 fills and 12311 sums of 189 items)